Add a helper script to create and template new blog posts
It's quite cumbersome to create the filenameand the front matter
by hand each time.

So created a helper script to create en ampty blog post with
current date and some default frontmatter.

You can place a .author_meta file in your repo to define some
sane defaults for your front matter. This file is ignored because
it depends on you.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>