Use the right image url in frontmatter
Docusaurus uses cards such as "twitter:" and and "og:". The URL
in the image frontmatter will be used for "twitter:image" and
"og:image" cards as preview image, if you post the blog post's URL
at Twitter or Facebook.

So as solution you should not use the avatar URL in this field.

also use a free image from Pixabay:
https://pixabay.com/photos/blini-salmon-pancakes-3771036/

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>