Fixing the first scan doc
For novice users it is cumbersome and error prone if the command
and its output is in one copyable fenced code block. Also it is no
good practice to prefix commands in shell/bash code with $ sign
or such because you must remove them always manually.

So this patch splits the example commands and removes the prifexs
so the command is plain pasteable w/o any manual edit.

Also fixed minor gramar bug: Before and after slash no space is
allowed[1].

[1]: https://www.grammarly.com/blog/slash/

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>