Fixing function to copy files from main repo
The function failed to create the destination directory because we
did the creation inside the conditional block which is only executed
when it already exists.

This patch moves the directory create out of the block. Also improves
logging a little bit: Writing waht is done before the action is better
because you have the log statement also if the executed action fails
and exit the code.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>