Pull values.yaml documentation to SCBv3
The current version of the documentation for the values.yaml file still
uses SCBv2 syntax, which is no longer used (see UPGRADING.md in the
main repo). I have updated this to the best of my knowledge, but someone
more familiar with the syntax should take another look at this.
In particular, I am unsure if all fields can indeed be set for both
scanner and parser images, or if some are exclusive to one or the other.

Signed-off-by: Max Maass <max.maass@iteratec.com>