Add some words about the OWASP
We want to outline more that we are a OWASP project. This is first
step. Later we will add a more prominent section with OWASP logo.

This text is not added into landingpageContent.js because it is
pure static content which is used only once in this place. So there
is no need for abstracting the content way ino a configuration.

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>