Add a basic example for the runtime view
HEre we use a PlantUML generated diagram instead of Draw.io because
painting such a diagram with mouse is massive pain: If you need to add
or remove life-lines you always move the other lines back and forth.
This is tedious and time consuming for more than two participant.

In PlantUML this is just inserting/removing the lines in text and
the tedious layouting is done automatically.

The sequence is derived from the picture on htis blog post:
https://www.securecodebox.io/blog/2021/07/20/the-architecture-of-securecodebox-v2

Signed-off-by: Sven Strittmatter <sven.strittmatter@iteratec.com>